tests/integration-tests/requirements.txt
argparse
assertpy
aws-parallelcluster
boto3
fabric
jinja2
junitparser
pykwalify
pytest
pytest-datadir
pytest-html
pytest-rerunfailures
pytest-sugar
pytest-xdist
PyYAML
retrying
troposphere
untangle
wrapt_timeout_decorator
